requirements.txt
django~=2.2
psycopg2-binary
django-adminlte3
django-import-export==2.2.0
